Added experimental script to bypass SE limitation
An experimental script was added that transforms input SE reads or PE reads with insufficient overlap into the input accepted by Snowball (i.e. PE reads with insert size equal to the read length in this settings). The assembler then runs as if the input reads were SE. Disadvantage: The amount of the input data will double. Disclaimer: The algorithm was never benchmarked in this settings. Note: This is just an experimental solution.